working on bugfix
