+++ project_source_code/src/resources/js/friend_swipe_script.js
@@ -89,7 +89,7 @@ document.addEventListener('DOMContentLoaded', () => {
         });
     } else {
         console.error("[SwipeDebug] Right swipe button not found!");
-    }
+    } 
 
 
     // 🔽 Toggle view between profile details and user posts (same as original)
