+++ project_source_code/src/views/pages/swipe.hbs
@@ -31,13 +31,8 @@
     <div id="match-modal" class="modal">
         <div class="modal-content">
             <h2>It's a match!</h2>
-<<<<<<< HEAD
             <p>You and {{user.display_name}} have liked each other.</p>
-            <button id="btn btn-primary">Continue Swiping</button>
-=======
-            <p>You and {{swipeTarget.display_name}} have liked each other.</p>
-            <button id="close-match-modal">Continue Swiping</button>
->>>>>>> 7828401c870eafbc5871ff3fae554d78bbbea542
+            <button id="btn btn-primary close-match-modal">Continue Swiping</button>
         </div>
     </div>
     {{/if}}
